Distinguish ref and upstream_ref in event dispatch
These two should be separate parameters. As far as a module workflow
invokation we will use the same value for both - `upstream_ref`.
But for the event that is triggered as a callback after a downstream
build completes we want to call the `main` branch in the `ci` repository
with an `upstream_ref` parameter holding the branch that was used on the
module.