Migrate from `master` to `main` branch
Use of the `master` and `slave` terms in the computer industry is
widespread, but has been recently discouraged, as it was brought to
attention that the terms may be considered offensive.
GitHub has already moved their default repositories from the master
to main in many places. We're planning to do the same for our
repositories, this however requires some changes in the code.
The `run-workflow` repository was already using `main` branch, but the
code in that repo depends on the code of the `ci` repo, which was using
`master` branch. Once we migrate there from `master` to `main`, the
changes proposed in this commit will need to be applied to
`run-workflow` repo. However, this is not everything - dependencies to
the `ci` will also need to be updated and commited (see README).