Migrate from `master` to `main` branch
Use of the `master` and `slave` terms in the computer industry is
widespread, but has been recently discouraged, as it was brought to
attention that the terms may be considered offensive.
GitHub has already moved their default repositories from the `master`
to `main` in many places. We're planning to do the same for our
repositories, this however requires some changes in the code.
The `load-env-variables` repository was already using `main` branch,
but referenced 'master' in coule of places in the code. This commit
modifies those mentions (corresponding changes has been also
introduced in other repositories).