Move custom GHA actions to `ci` repository
We are moving custom actions used in the Keep/tBTC Continuous
Integration process from separate repositories to the `ci` repository.
This is a step that brings us closer to simplification of the process of
updating CI configs. Curently the moved actions still need to be rebuilt
after changes to the `ci/config/config.json/', but eventually we want to
handle the dependencies differently and for that we need to have the
config and actions in the same repository.