Stabilize failing test
The test checking if `npm-version-bump` action "updates version for
already published environment that don't match initial preid" was
unstable - was refering to actively worked on npm packages versions,
making it necessary to update the action every time new package with
that base version is released on `ropsten`.
In this commit we change the reference to different version, which is no
longer actively worked on.