Incorporate workflow building keep-ecdsa initcontainer
In `keep-ecdsa` repository we're changing the kube configuration for
keep-ecdsa pods, so that the population of TBTCSystem address would
happen automatically. For the feature to work correctly we are
separating the job building/publishing the initcontainer from the
workflow migrating Solidity contracts. This way we will be able to run
initcontainer job after migrating tbtc condtracts.
Now we need reflect that order in the `./config/config.json` file.