Update meercode links
We've been using Meercode tool to gather information about GH Actions
workflows triggered by a cron and by workflow_dispatch events, which
has been useful during our daily examination of tests run at night or
during Ropsten deployment.
There was a bug in meercode that was limitig functionality of lists
filtering, which forced us to have non-optimal configuration of the
views. Now the bug got fixed and we are able to present all workflows
with the daily test in one view.