Incorporate workflow building t-network solidity-contracts
We're adding a new module (`threshold-network/solidity-contracts`) to
the CI flow configuration. As the module should be built before
`tbtc-v2` and `coverage-pools` modules and will be changed more often
than `tbtc`, it's placed between `tbtc` and `tbtc-v2` in the CI flow.
Apart from changing the `config.json file, we're also updating actions
which rely on this file.