Replace babel with @vercel/ncc
We want to simplify actions code generation to reduce build times and
complexity. Solution recommended by GitHub documentation is using @vercel/ncc
https://docs.github.com/en/actions/creating-actions/creating-a-javascript-action#commit-tag-and-push-your-action-to-github
We tried it before but for some reason the tool wasn't working well with
our code, so we switched to babel. It turns out that the latest version
of ncc is able to handle our code generation, hence we're switching to
it.

The change let us simplify also linting configuration to have just once
config.