Create a config file for CI flow in V2
In V2 of our system the CI flow will look different than in V1.
There are different dependencies between projects, so we need to adjust
the config file to build the right modules in the right order. For now
the order is:
1. threshold-network/solidity-contracts
2. keep-network/keep-core/solidity/random-beacon
3. keep-network/keep-core/solidity/ecdsa
4. keep-network/tbtc-v2/solidity
In the future we'll also incorporate building of other modules, like
client code or dashboards.

We're moving the old V1 configuration to config-v1.json file, in case it
will be needed someday.

Note: Apart from the change bringing new V2 configuration, we also nedd
to re-install code of specific actions using the config. This will be
done in separate commit.