Use `v2` of `run-workflow` action in the Main workflow
In PR #28 we modified the order in which modules get executed as part of
the CI flow. New modules were added, old rmoved, etc. We tagged the new
config with `v2` tag, which results in need for calling `v2`-tagged
actions in our CI. We already updated the workflows building each
module in previous code changes, but we did not update the `Main`
workflow in the `ci` project - and this workflow is responsible for
triggering new module once the previous (upstream) one finishes
execution. Now we're adding that missing update.