Update CI actions using Docker script
When the config of the CI flow changes, we need to update the
`notify-workflow-completed`, `run-workflow` and `upstream-builds-query` actions
so that they would use the new config. This can be done separately for each
action (steps are described in the action's READMEs). Alternatively, this can be
done via the script I provide in this PR. The exact steps are described in the
comments in the script.