Change authentication method during `git clone`
With the previous method, after executing `hub pull-request...` command, `hub`
was prompting to provide username and password. Now that we've renamed the API
TOKEN secret to `GITHUB_TOKEN`, `hub` recognizes it and does not prompt for the
usrname/password. The change has been verified to work, we've succesfully
created a PR using this config.