Set Node.js version to 18.15.0
We want to fix problem with the `hardhat compile` command excecuted during
building Markdow Docs sometimes not generating expected artifacts. The problem
is caused by the process silently quitting, which is related to the used version
of Node (as described in NomicFoundation/hardhat#3877). We're setting the
version to a fixed `v18.15.0` value, as this is the version that helped with the
similar problems in other projects.