Remove `typescript` (and `optimistic-minting`) modules from the CI flow
We no longer want to publish new `typescript` testnet packages as part of the CI
flow. The `typescript` project is transitioning to v2, which has all the chain
configs embeded, so we'll no longer use the `mainnet` and `goerli` packages.
All existing testnet packages were marked in the NPM registry as deprecated
(projects still can use them as dependencies, but will get a warning).